data_prep/yelp_data_scraping/requirements.txt
python-dotenv
pandas
requests
beautifulsoup4
pdfplumber
